Fix incorrect format placeholder
Not only did the format placeholder not match the variable, the
variable also didn't match the function it was getting its value from.

+++ src/backend/backup/basebackup_incremental.c
@@ -428,7 +428,7 @@ PrepareForIncrementalBackup(IncrementalBackupInfo *ib,
 	while (1)
 	{
 		long		timeout_in_ms = 10000;
-		unsigned	elapsed_seconds;
+		long		elapsed_seconds;
 
 		/*
 		 * Align the wait time to prevent drift. This doesn't really matter,
@@ -493,7 +493,7 @@ PrepareForIncrementalBackup(IncrementalBackupInfo *ib,
 			TimestampDifferenceMilliseconds(initial_time, current_time) / 1000;
 		ereport(WARNING,
 				(errcode(ERRCODE_OBJECT_NOT_IN_PREREQUISITE_STATE),
-				 errmsg("still waiting for WAL summarization through %X/%X after %d seconds",
+				 errmsg("still waiting for WAL summarization through %X/%X after %ld seconds",
 						LSN_FORMAT_ARGS(backup_state->startpoint),
 						elapsed_seconds),
 				 errdetail("Summarization has reached %X/%X on disk and %X/%X in memory.",
